+++ lib/src/icon_data.dart
@@ -1,122 +1,67 @@
 import 'package:flutter/widgets.dart';
 
-/// [IconData] for a font awesome brand icon from a code point
+/// A wrapper for [IconData] to ensure type safety and correct rendering
+/// with [FaIcon].
 ///
-/// Code points can be obtained from fontawesome.com
-class IconDataBrands extends IconData {
-  const IconDataBrands(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeBrands',
-        fontPackage: 'font_awesome_flutter',
-      );
+/// This class prevents usage of Font Awesome icons in the standard [Icon] widget,
+/// which would result in incorrect rendering (clipping/misalignment) for
+/// non-square icons.
+@immutable
+final class FaIconData {
+  /// The underlying [IconData] for this font awesome icon.
+  final IconData data;
+
+  /// Creates a wrapper for the given [data].
+  const FaIconData(this.data);
+
+  /// The code point of the icon.
+  int get codePoint => data.codePoint;
+
+  /// The font family of the icon.
+  String? get fontFamily => data.fontFamily;
+
+  /// The font package of the icon.
+  String? get fontPackage => data.fontPackage;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other is FaIconData && other.data == data;
+  }
+
+  @override
+  int get hashCode => data.hashCode;
+
+  @override
+  String toString() => 'FaIconData(data: $data)';
 }
 
-/// [IconData] for a font awesome solid icon from a code point
+/// [FaIconData] for a font awesome duotone icon.
 ///
-/// Code points can be obtained from fontawesome.com
-class IconDataSolid extends IconData {
-  const IconDataSolid(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeSolid',
-        fontPackage: 'font_awesome_flutter',
-      );
-}
+/// Duotone icons consist of a primary and a secondary layer.
+@immutable
+final class FaDuotoneIconData {
+  /// The primary layer of the icon.
+  final IconData primary;
 
-/// [IconData] for a font awesome regular icon from a code point
-///
-/// Code points can be obtained from fontawesome.com
-class IconDataRegular extends IconData {
-  const IconDataRegular(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeRegular',
-        fontPackage: 'font_awesome_flutter',
-      );
-}
+  /// The secondary layer of the icon.
+  final IconData secondary;
 
-/// [IconData] for a font awesome light icon from a code point. Only works if
-/// light icons (font awesome pro) have been installed.
-///
-/// Code points can be obtained from fontawesome.com
-class IconDataLight extends IconData {
-  const IconDataLight(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeLight',
-        fontPackage: 'font_awesome_flutter',
-      );
-}
+  /// Creates a duotone icon data.
+  const FaDuotoneIconData({required this.primary, required this.secondary});
 
-/// [IconData] for a font awesome duotone icon from a code point. Only works if
-/// duotone icons (font awesome pro) have been installed.
-///
-/// Code points can be obtained from fontawesome.com. Each duotone icon consists
-/// of a primary [codePoint] and a [secondary].
-class IconDataDuotone extends IconData {
-  /// Secondary glyph of the duotone icon
-  ///
-  /// Due to tree-shaking restraints [secondary] cannot be the codepoint itself,
-  /// but has to be an [IconData] object.
-  final IconData? secondary;
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other is FaDuotoneIconData &&
+        other.primary == primary &&
+        other.secondary == secondary;
+  }
 
-  const IconDataDuotone(super.codePoint, {this.secondary})
-    : super(
-        fontFamily: 'FontAwesomeDuotone',
-        fontPackage: 'font_awesome_flutter',
-      );
-}
+  @override
+  int get hashCode => Object.hash(primary, secondary);
 
-/// [IconData] for a font awesome thin icon from a code point. Only works if
-/// thin icons (font awesome pro, v6+) have been installed.
-///
-/// Code points can be obtained from fontawesome.com
-class IconDataThin extends IconData {
-  const IconDataThin(super.codePoint)
-    : super(fontFamily: 'FontAwesomeThin', fontPackage: 'font_awesome_flutter');
-}
-
-/// [IconData] for a font awesome sharp thin icon from a code point. Only works if
-/// thin icons (font awesome pro, v6+) have been installed.
-///
-/// Code points can be obtained from fontawesome.com
-class IconDataSharpThin extends IconData {
-  const IconDataSharpThin(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeSharpThin',
-        fontPackage: 'font_awesome_flutter',
-      );
-}
-
-/// [IconData] for a font awesome sharp light icon from a code point. Only works if
-/// thin icons (font awesome pro, v6+) have been installed.
-///
-/// Code points can be obtained from fontawesome.com
-class IconDataSharpLight extends IconData {
-  const IconDataSharpLight(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeSharpLight',
-        fontPackage: 'font_awesome_flutter',
-      );
-}
-
-/// [IconData] for a font awesome sharp regular icon from a code point. Only works if
-/// thin icons (font awesome pro, v6+) have been installed.
-///
-/// Code points can be obtained from fontawesome.com
-class IconDataSharpRegular extends IconData {
-  const IconDataSharpRegular(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeSharpRegular',
-        fontPackage: 'font_awesome_flutter',
-      );
-}
-
-/// [IconData] for a font awesome sharp solid icon from a code point. Only works if
-/// thin icons (font awesome pro, v6+) have been installed.
-///
-/// Code points can be obtained from fontawesome.com
-class IconDataSharpSolid extends IconData {
-  const IconDataSharpSolid(super.codePoint)
-    : super(
-        fontFamily: 'FontAwesomeSharpSolid',
-        fontPackage: 'font_awesome_flutter',
-      );
+  @override
+  String toString() =>
+      'FaDuotoneIconData(primary: $primary, secondary: $secondary)';
 }